Started refactoring
Encapsulated args parsing and post-processing. Moved most of init.py code to main method to keep functionality while moving methods around.
